% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calc_completeness_rust}
\alias{calc_completeness_rust}
\title{Completeness function to calculate the completeness at given positions.}
\usage{
calc_completeness_rust(
  ra_observed,
  dec_observed,
  ra_target,
  dec_target,
  angular_radius
)
}
\arguments{
\item{ra_observed}{The Right Ascension of the galaxies that were observed, in degrees.}

\item{dec_observed}{The Declination of the galaxies in degrees that were observed, in degrees.}

\item{ra_target}{The Right Ascension of the galaxies that were targeted for observation in degrees.}

\item{dec_target}{The Declination of the galaxies that were targeted for observation in degress.}

\item{angular_radius}{The search area for each galaxy in degress.}
}
\value{
An array of elements between 0-1.
}
\description{
Completeness function to calculate the completeness at given positions.
}
