% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{calculate_completeness}
\alias{calculate_completeness}
\title{Calculating the completeness of a survey.}
\usage{
calculate_completeness(
  ra_observed,
  dec_observed,
  ra_target,
  dec_target,
  ra_eval,
  dec_eval,
  search_radii
)
}
\arguments{
\item{ra_observed}{The Right Ascension of the galaxies that were observed, in degrees.}

\item{dec_observed}{The Declination of the galaxies in degrees that were observed, in degrees.}

\item{ra_target}{The Right Ascension of the galaxies that were targeted for observation in degrees.}

\item{dec_target}{The Declination of the galaxies that were targeted for observation in degress.}

\item{search_radii}{The angular area for each eval point to search for members in degrees.}

\item{ra_evaulate}{The Right Ascension to be evaluated for completeness in degrees.}

\item{dec_evaluate}{The Declination to be evaluated for completeness in degrees.}
}
\value{
An array equal to the length of the eval ra and decs with every element between 0-1.
}
\description{
\verb{calculate completeness} uses the positions of the galaxies that were observed out of the full
target list and evaluates the completeness at given positions within some given area.
}
\details{
The completeness of a survey can be evaluated at any given (RA, Dec) position if the catalog of
galaxies that were planned to be observed (the target catalog) and the actual final observed
catalog of galaxies (the observed catalog). This function simply takes the positions that need
to be evaluated and counts the number of galaxies in the observed and target catalogs within some
given angular radius., The completeness is simply the ratio of these two counts, resulting in a
number between 0-1. This can then be used by Nessie to adjust the linking lengths in areas of
lower completeness.
}
