% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calculate_cost_metrics}
\alias{calculate_cost_metrics}
\title{Calculating the cost metrics for comparison to mock catalogs.}
\usage{
calculate_cost_metrics(measured_groups, mock_groups, singleton_id)
}
\arguments{
\item{measured_groups}{Array of the group ids as an integer of the measured groups.}

\item{mock_groups}{Array of the mock group ids as integers.}

\item{singleton_ids}{the id value that is assigned to the singleton cases (groups of one member). Should be an integer (Usually negative).}
}
\value{
A list of the values of equations 9 - 15 in Robotham+2011.
}
\description{
This function will compare the measured groups to ones of known grouping (usually from simulations.)
This is done by calculating the metrics defined as equations (9 - 15) in Robotham+2011.
}
