% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calculate_max_sigmas}
\alias{calculate_max_sigmas}
\title{Calculate the Sigma from a given mass for a range of redshift values.}
\usage{
calculate_max_sigmas(
  max_solar_mass,
  redshift_array,
  omega_m,
  omega_k,
  omega_l,
  h0
)
}
\arguments{
\item{max_solar_mass}{The maximum viral mass in solar masses for the viral radii.}

\item{redshift_array}{an array of multiple redshift values.}

\item{omega_m}{Mass density (often 0.3 in LCDM).}

\item{omega_k}{Effective mass density of relativistic particles (often 0. in LCDM).}

\item{omega_l}{Effective mass density of dark energy (often 0.7 in LCDM).}

\item{hubble_constant}{H0 = 100 * h.}
}
\description{
Calculate the Sigma from a given mass for a range of redshift values.
}
