% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{calculate_s_score}
\alias{calculate_s_score}
\title{Determines the s score as in Robotham+2011}
\usage{
calculate_s_score(measured_groups, mock_groups, min_group_size)
}
\description{
Determines the s score as in Robotham+2011
}
