% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{calculate_s_total}
\alias{calculate_s_total}
\title{Comparing measured groups to mock catalogue groups}
\usage{
calculate_s_total(measured_ids, group_ids, min_group_size = 2)
}
\arguments{
\item{mock_group_ids}{group ids from the mock catalog. Have to be integers.}

\item{singleton_ids}{the integer id value that is assigned to all singleton galaxies.}
}
\value{
a list containing all the values from equations 9 - 15 in Robotham+2011
}
\description{
Compares the current group_ids to a mock known grouping ids.
}
