% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{create_density_function}
\alias{create_density_function}
\title{Running density function estimation}
\usage{
create_density_function(
  redshifts,
  total_counts,
  survey_fractional_area,
  cosmology,
  binwidth = 40,
  N = 10000
)
}
\arguments{
\item{redshifts}{An array of redshift values.}

\item{total_counts}{The total counts of the redshift redshift survey. Doesn't make sense.}

\item{survey_fractional_area}{The fraction of the survey over a full 4pi steradian area.}

\item{cosmology}{A FlatCosmology object (created using the FlatCosmology class).}

\item{binwidth}{The binwidth (in comoving distance) which will be used for the integration}

\item{N}{The total number of data points to interpolate over.}
}
\value{
running_density A function which is the rho(z).
}
\description{
\code{create_density_function} returns the rho(z) function (Eq ?? in Lambert+2025).
This is a required argument for the FoF algorithm
}
\details{
The n(z) is approximated from a given array of redshift values. These values can be generated
in any way. In the simpliest case this can just be the redshifts of the redshift survey.
More approriate methods should be used though. The Probability Distribution function is then
estimated and integrated over several bins. The total running density is simply the total counts
in a bin multipled by the PDF and divided by the running volume.
}
