% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{create_flat_cosmology}
\alias{create_flat_cosmology}
\title{Create a Flat LCDM cosmology parameter object.}
\usage{
create_flat_cosmology(h, omega_m)
}
\arguments{
\item{h}{Little h. Often taken to be 0.7. H0 = 100 * h.}

\item{omega_m}{Density of matter in the universe. Usually taken to be 0.3}
}
\value{
parameters A named list with the cosmological parameters.
}
\description{
\code{create_flat_cosmology} will return a named list with the cosmological parameters
with OmL = 1 - Om0.
}
\examples{
cosmo <- create_cosmology(H0 = 70, Om0 = 0.3)
cosmo$H0  # Access via $
}
