% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{create_group_catalog}
\alias{create_group_catalog}
\title{Creates a group catalog from the given arrays.}
\usage{
create_group_catalog(
  ra,
  dec,
  redshift,
  absolute_magnitudes,
  velocity_errors,
  group_ids,
  omega_m,
  omega_k,
  omega_l,
  h0
)
}
\arguments{
\item{ra}{Array of right asscension values.}

\item{dec}{Array of declination values.}

\item{redshift}{Array of redshift values.}

\item{velocity_errors}{Array of velocity errors.}

\item{magnitudes}{Array of apparent magnitude values.}

\item{group_id}{Array of the group ids, where -1 represents galaxies not in a group.}
}
\value{
A named list with ra, dec, redshift, co_dist, r50, r100, rsigma, and multiplicity.
}
\description{
This is the R wrapper for the rust functionality that builds the group catalog. There shouldn't
be a need to use this in R as a more R-friendly function should be available.
}
