% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{create_pair_catalog}
\alias{create_pair_catalog}
\title{Creates a pair catalog from the given arrays.}
\usage{
create_pair_catalog(ra, dec, redshift, absolute_magnitudes, group_ids)
}
\arguments{
\item{ra}{Array of right asscension values.}

\item{dec}{Array of declination values.}

\item{redshift}{Array of redshift values.}

\item{magnitudes}{Array of apparent magnitude values.}

\item{group_id}{Array of the group ids, where -1 represents galaxies not in a group.}
}
\value{
A named list with pair properties.
}
\description{
This is the R wrapper for the rust functionality that builds the pair catalog. There shouldn't
be a need to use this in R as a more R-friendly function should be available.
}
