% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_funcs.R
\name{.group_graph}
\alias{.group_graph}
\title{Finding groups from a graph structure.}
\usage{
.group_graph(links)
}
\arguments{
\item{links}{A named list with i and j. These links are made using the \code{fof_links} function.}
}
\value{
A dataframe with galaxy ids in one column and group id in another.
}
\description{
Creates a graph structure from the galaxy pairs and then finds the groups within those pairs.
}
\details{
The pairs of galaxies can be though of as edges of a graph. An undirected graph data structure
can be built from the edges. All the connected components of the graph would then be the groups.
These are returned as such.

\code{.group_graph} should not be used by the user. This is a helper function for the fof function.
}
