% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmology.R
\name{FlatCosmology}
\alias{FlatCosmology}
\title{FlatCosmology class}
\description{
An R6 class for flat ΛCDM cosmology models, providing methods for computing
comoving distances, virial radii, velocity dispersions, and Hubble evolution.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{h}}{Dimensionless Hubble parameter (i.e., H0 / 100).}

\item{\code{hubble_constant}}{Hubble constant in km/s/Mpc.}

\item{\code{omega_m}}{Matter density parameter (Ωₘ).}

\item{\code{omega_lambda}}{Cosmological constant / dark energy density (Ω_Λ), derived assuming flatness.}

\item{\code{omega_k}}{Curvature density parameter (Ωₖ), fixed to 0 in flat cosmology.}

\item{\code{omega_radiation}}{Radiation density parameter (Ωᵣ), fixed to 0 by default.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FlatCosmology-new}{\code{FlatCosmology$new()}}
\item \href{#method-FlatCosmology-comoving_distance}{\code{FlatCosmology$comoving_distance()}}
\item \href{#method-FlatCosmology-dist_mod}{\code{FlatCosmology$dist_mod()}}
\item \href{#method-FlatCosmology-z_at_comoving_dist}{\code{FlatCosmology$z_at_comoving_dist()}}
\item \href{#method-FlatCosmology-virial_radius}{\code{FlatCosmology$virial_radius()}}
\item \href{#method-FlatCosmology-velocity_dispersion}{\code{FlatCosmology$velocity_dispersion()}}
\item \href{#method-FlatCosmology-h0_grow}{\code{FlatCosmology$h0_grow()}}
\item \href{#method-FlatCosmology-differential_covol}{\code{FlatCosmology$differential_covol()}}
\item \href{#method-FlatCosmology-clone}{\code{FlatCosmology$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-new"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-new}{}}}
\subsection{Method \code{new()}}{
Create a new FlatCosmology object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$new(h, omega_matter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h}}{Dimensionless Hubble parameter (H0 / 100).}

\item{\code{omega_matter}}{Matter density parameter Ωₘ.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-comoving_distance"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-comoving_distance}{}}}
\subsection{Method \code{comoving_distance()}}{
Compute comoving distances for a vector of redshifts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$comoving_distance(redshift)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{redshift}}{A numeric vector of redshifts.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of comoving distances (in Mpc).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-dist_mod"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-dist_mod}{}}}
\subsection{Method \code{dist_mod()}}{
Compute the distance modulus for a vector of redshifts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$dist_mod(redshift)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{redshift}}{A numeric vector of redshifts.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of distance moduli.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-z_at_comoving_dist"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-z_at_comoving_dist}{}}}
\subsection{Method \code{z_at_comoving_dist()}}{
Compute redshift values corresponding to comoving distances.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$z_at_comoving_dist(co_dist)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{co_dist}}{A numeric vector of comoving distances (in Mpc).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of redshifts.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-virial_radius"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-virial_radius}{}}}
\subsection{Method \code{virial_radius()}}{
Compute the virial radius of halos with given mass and redshift.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$virial_radius(mass_solar, redshift)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mass_solar}}{A numeric vector of halo masses in solar masses.}

\item{\code{redshift}}{A numeric vector of redshifts.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of virial radii (in kpc).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-velocity_dispersion"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-velocity_dispersion}{}}}
\subsection{Method \code{velocity_dispersion()}}{
Compute the velocity dispersion of halos with given mass and redshift.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$velocity_dispersion(mass_solar, redshift)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mass_solar}}{A numeric vector of halo masses in solar masses.}

\item{\code{redshift}}{A numeric vector of redshifts.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of velocity dispersions (in km/s).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-h0_grow"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-h0_grow}{}}}
\subsection{Method \code{h0_grow()}}{
Compute the redshift-dependent Hubble parameter H(z) in a flat cosmology.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$h0_grow(redshift)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{redshift}}{A numeric vector of redshifts.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of H(z) values (in km/s/Mpc).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-differential_covol"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-differential_covol}{}}}
\subsection{Method \code{differential_covol()}}{
Compute the redshift-dependent Hubble parameter H(z) in a flat cosmology.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$differential_covol(redshift)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{redshift}}{A numeric vector of redshifts.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of differential comoving volume in Mpc/sr.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FlatCosmology-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FlatCosmology-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FlatCosmology$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
