% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{fof}
\alias{fof}
\title{Run the friends-of-friends algorithm}
\usage{
fof(
  b0,
  r0,
  ras,
  decs,
  redshifts,
  density_function,
  cosmology,
  completeness = rep(1, length(ras)),
  Mmax = 1e+15
)
}
\arguments{
\item{b0}{Plane-of-sky constant which will be scaled by the given linking lengths.}

\item{r0}{Line-of_sight constant value which will be scaled by the given linking lengths and b0.}

\item{ras}{An array-like object of right ascension in decimal degrees.}

\item{decs}{An array-like object of declination values in decimal degrees.}

\item{redshifts}{An array-like object of comoving_distances in Mpc.}

\item{density_function}{An array-like object of linking lengths for every galaxy. These are before}

\item{cosmology}{A cosmology object. (can be created with \code{create_flat_cosmology}).}

\item{completeness}{An array of equal size to the data with values between 0 and 1. Representing how
complete the survey is around that particular galaxy. This will scale that galaxy's linking lengths
accordingly.}
}
\value{
Data Frame of the galaxy ids and which groups they are in.
}
\description{
Performs the friends-of-friends algorithm on the given data.
}
\details{
This is the core function of the group finder and actually runs the friends-of-friends algorithm.
Before running this function it is important to have the density function already calculated.
This can be done by passing an array of n(z) values to the \code{density_function}. An array of completeness
can also be passed if needed or else 100\% completeness is assumed.
}
