% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fof_links}
\alias{fof_links}
\title{finding the links between all galaxies}
\usage{
fof_links(ra_array, dec_array, comoving_distances, linking_lengths, b0, r0)
}
\arguments{
\item{comoving_distances}{Array of comoving distances in Mpc.}

\item{linking_lengths}{An array of individual scaled linking lengths for each galaxy (ignoring r0 and b0).}

\item{b0}{The plane-of-sky constant to be scaled.}

\item{r0}{The line-of-sight constant to be scaled.}

\item{ra}{Array of right ascension values.}

\item{dec}{Array of declination values.}
}
\value{
A dataframe-like object of tuples which represent the link between galaxies (i, j) if they exist.
}
\description{
\code{fof_links} will determine all connections between galaxies in a survey and return the pairs.
}
