% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{gen_random_redshifts}
\alias{gen_random_redshifts}
\title{Creating randoms catalogue to model the n(z) without underlying large-scale structure}
\usage{
gen_random_redshifts(
  redshifts,
  mags,
  z_lim,
  maglim,
  n_clone,
  iterations,
  cosmo
)
}
\arguments{
\item{redshifts}{Redshifts from the survey.}

\item{mags}{Magnitudes from the survey.}

\item{z_lim}{The maximum z over which to confine all calculations.}

\item{maglim}{The magnitude limit of the survey.}

\item{n_clone}{The number of times more the randoms catalogue should be}

\item{iterations}{The number of iterations to iteratively iterate. (5-10 should be plenty)}

\item{cosmology}{A FlatCosmology object (created using the FlatCosmology class).}
}
\value{
An array of redshift values roughly n_clone times as many as the catalogue.
}
\description{
\code{gen_random_redshifts} creates a randoms catalogue following the Farrow+2015 method. These random
redshifts accuratly model the n(z) distribution without the underylying large-scale structure.
These can then be saved or used with the \code{create_density_function} to build the density function
needed for the group finder.
}
