% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_properties.R
\name{generate_group_catalog}
\alias{generate_group_catalog}
\title{Generating the group catalog}
\usage{
generate_group_catalog(
  ra,
  dec,
  redshift,
  absolute_magnitudes,
  velocity_errors,
  group_ids,
  cosmology
)
}
\arguments{
\item{ra}{The right ascension of the entire redshift catalog.}

\item{dec}{The declination of the entire redshift catalog.}

\item{redshift}{The redshifts of the entire redshift catalog.}

\item{group_ids}{The IDs of the groups for every galaxy in the redshift catalog.}

\item{cosmology}{a Flat LCDM cosmology object. This can be generated with \code{create_flat_cosmology}}

\item{magnidues}{The magnitudes for the entire redshift catalog.}

\item{velocity_erros}{The velocity errors for the entire redshift catalog.}
}
\value{
Named list with group centers, radii, and multiplicity
}
\description{
Generating the group catalog
}
