% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_group_ids}
\alias{get_group_ids}
\title{Get the group ids for the given redshift survey.}
\usage{
get_group_ids(
  b0,
  r0,
  ras,
  decs,
  redshifts,
  density_function,
  cosmology,
  completeness = rep(1, length(ras))
)
}
\arguments{
\item{b0}{Plane-of-sky constant which will be scaled by the given linking lengths.}

\item{r0}{Line-of_sight constant value which will be scaled by the given linking lengths and b0.}

\item{ras}{An array-like object of right ascension in decimal degrees.}

\item{decs}{An array-like object of declination values in decimal degrees.}

\item{redshifts}{An array-like object of comoving_distances in Mpc.}

\item{density_function}{An array-like object of linking lengths for every galaxy. These are before}

\item{cosmology}{A cosmology object. (can be created with \code{create_flat_cosmology}).}

\item{completeness}{An array of equal size to the data with values between 0 and 1. Representing how
complete the survey is around that particular galaxy. This will scale that galaxy's linking lengths
accordingly.}
}
\value{
A single array of group ids equal to the length of the arrays that were given.
}
\description{
Performs the FoF algorithm on the given data and returns the group_id column for that redshift survey.
}
\details{
This function is a helper function which wraps around the \code{fof} function but instead of returning
the galaxy_id and associated group, this function will return a single array of group_ids which is
equal to the length of the arrays. The ids here do not differentiate between binaries and groups.
Galaxies which were not found in a group are given the -1 group id.

This is built to work well with the \code{generate_group_catalog} function and will put the group ids
in the order that is required for that function.
}
