% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{RedshiftCatalog}
\alias{RedshiftCatalog}
\title{RedshiftCatalog class}
\description{
A catalog object for redshift-space galaxy data that supports group finding
using a Friends-of-Friends (FoF) algorithm with redshift-dependent linking lengths.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ra_array}}{A numeric vector of Right Ascension values (in degrees).}

\item{\code{dec_array}}{A numeric vector of Declination values (in degrees).}

\item{\code{redshift_array}}{A numeric vector of redshift values.}

\item{\code{density_function}}{A function that takes redshift values and returns the local density.}

\item{\code{cosmology}}{A FlatCosmology object (created using the FlatCosmology class).}

\item{\code{completeness}}{A numeric vector of completeness values per galaxy (default is 1 for all).}

\item{\code{group_ids}}{An integer vector storing the resulting group ID for each galaxy (after \code{run_fof()} is called).}

\item{\code{current_r0}}{The most recent \code{r0} used in \code{run_fof()}}

\item{\code{current_b0}}{The most recent \code{b0} used in \code{run_fof()}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RedshiftCatalog-new}{\code{RedshiftCatalog$new()}}
\item \href{#method-RedshiftCatalog-get_raw_groups}{\code{RedshiftCatalog$get_raw_groups()}}
\item \href{#method-RedshiftCatalog-run_fof}{\code{RedshiftCatalog$run_fof()}}
\item \href{#method-RedshiftCatalog-calculate_group_table}{\code{RedshiftCatalog$calculate_group_table()}}
\item \href{#method-RedshiftCatalog-clone}{\code{RedshiftCatalog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RedshiftCatalog-new"></a>}}
\if{latex}{\out{\hypertarget{method-RedshiftCatalog-new}{}}}
\subsection{Method \code{new()}}{
Create a new RedshiftCatalog object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftCatalog$new(
  ra_array,
  dec_array,
  redshift_array,
  density_function,
  cosmology,
  completeness = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ra_array}}{A numeric vector of right ascension values (degrees).}

\item{\code{dec_array}}{A numeric vector of declination values (degrees).}

\item{\code{redshift_array}}{A numeric vector of redshifts.}

\item{\code{density_function}}{A function that takes redshifts and returns galaxy density.}

\item{\code{cosmology}}{A FlatCosmology object (created using the FlatCosmology class).}

\item{\code{completeness}}{An optional numeric vector of completeness weights (default: 1 for all galaxies).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RedshiftCatalog-get_raw_groups"></a>}}
\if{latex}{\out{\hypertarget{method-RedshiftCatalog-get_raw_groups}{}}}
\subsection{Method \code{get_raw_groups()}}{
Generate FoF links between galaxies based on spatial and redshift linking lengths. There is
very little reason to have to run this yourself. In most cases it is more appropriate to run
the \code{run_fof} method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftCatalog$get_raw_groups(b0, r0, max_stellar_mass = 1e+15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b0}}{The plane-of-sky linking length constant.}

\item{\code{r0}}{The line-of-sight linking length constant.}

\item{\code{max_stellar_mass}}{The maximum stellar mass to cap linking distances (default: 1e15 solar masses).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame with columns \code{galaxy_id} and \code{group_id} indicating linked galaxies.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RedshiftCatalog-run_fof"></a>}}
\if{latex}{\out{\hypertarget{method-RedshiftCatalog-run_fof}{}}}
\subsection{Method \code{run_fof()}}{
Run the full Friends-of-Friends (FoF) algorithm and assign group IDs to all galaxies.
Singleton galaxies (unlinked) are given group ID -1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftCatalog$run_fof(b0, r0, max_stellar_mass = 1e+15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b0}}{The plane-of-sky linking length scale factor.}

\item{\code{r0}}{The line-of-sight linking length scale factor.}

\item{\code{max_stellar_mass}}{The maximum stellar mass to cap linking distances (default: 1e15).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RedshiftCatalog-calculate_group_table"></a>}}
\if{latex}{\out{\hypertarget{method-RedshiftCatalog-calculate_group_table}{}}}
\subsection{Method \code{calculate_group_table()}}{
Generate a summary data.frame of group properties based on assigned group IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftCatalog$calculate_group_table(absolute_magnitudes, velocity_errors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{absolute_magnitudes}}{A numeric vector of absolute magnitudes per galaxy.}

\item{\code{velocity_errors}}{A numeric vector of redshift or velocity errors.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame summarizing group-level statistics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RedshiftCatalog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RedshiftCatalog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RedshiftCatalog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
