% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{tune_group_finder}
\alias{tune_group_finder}
\title{Tuning the group finder to a mock catalogue to find the optimum b0 and r0.}
\usage{
tune_group_finder(
  list_of_catalogs,
  minimum_group_size,
  b0_estimate,
  r0_estimate,
  b0_bounds,
  r0_bounds,
  n_iterations = c(100, 100),
  nfinal_mcmc = 2500
)
}
\arguments{
\item{list_of_catalogs}{A list of RedshiftCatalog objects with mock_group_ids already set.}

\item{minimum_group_size}{The minim size of groups that should be accounted for when tuning.}

\item{b0_estimate}{The initial guess of the b0 linking length}

\item{r0_estimate}{The initial fiess of the r0 linking length}

\item{b0_bounds}{The lower and upper limits to search for the b0 linking length (e.g. c(0.03, 0.1)).}

\item{r0_bounds}{The lower and upper limits to search for the r0 linking length (e.g. c(5, 50)).}

\item{n_iterations}{The Number of iterations per CMA and Laplace's Demon respectively. Set to (100, 100)}

\item{n_final_mcmc}{The Number of iterations to run for the final MCMC run. Set to 2500}
}
\value{
A list of output from \code{Highlander} containg at least the best parameters of all iterations. See \code{Highlander} for more details.
}
\description{
Uses Highlander(CMA+MCMC fitting) to find the optimum parameters for b0 and r0 given a mock catalogue
or list of mock catalgoues.
}
\details{
This function will average over all lightcones that are passed to it making it easy to pass
multiple independent lightcones which some users may wish to do to account for cosmic variance.
At the same time, the user can pass a single lightcone if that is all they wish. This function
is already parralized in multiple different ways so the user should not try and parralize it themselves
as this can actually slow the process down significantly. It should scale easily with number of cores.

This is a helper function and it isn't necessary to use for tuning. Users can define the optimum
parameters in any way they wish. This is just a good approach that should meet most use cases.
}
