% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{z_at_comoving_distances}
\alias{z_at_comoving_distances}
\title{redshift at some given comoving distances in Mpc}
\usage{
z_at_comoving_distances(distances, omega_m, omega_k, omega_l, h0)
}
\description{
redshift at some given comoving distances in Mpc
}
